ref: move getReducer to own file
